% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.selection.R
\name{print.selection}
\alias{print.selection}
\title{Short \code{selection} summary}
\usage{
\method{print}{selection}(x = model, ...)
}
\arguments{
\item{x}{\code{selection} object.}

\item{\ldots}{Other options.}
}
\value{
The function returns the best subset of size q and its information
  criterion value. In the case of \code{seconds=TRUE} this information is
  returned for each alternative model.
}
\description{
\code{\link{selection}} summary
}
\examples{
library(FWDselect)
data(diabetes)
x = diabetes[ ,2:11]
y = diabetes[ ,1]
obj1 = selection(x, y, q = 1, method = "lm", criterion = "variance", cluster = FALSE)
obj1
}
\author{
Marta Sestelo, Nora M. Villanueva and Javier Roca-Pardinas.
}
\seealso{
\code{\link{selection}}.
}

