% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.qselection.R
\name{print.qselection}
\alias{print.qselection}
\title{Short \code{qselection} summary}
\usage{
\method{print}{qselection}(x = object, ...)
}
\arguments{
\item{x}{\code{qselection} object.}

\item{\ldots}{Other options.}
}
\value{
The function returns a summary table with the subsets of size
  \eqn{q}, their information criterion values and the chosen variables for
  each one.
}
\description{
\code{\link{qselection}} summary
}
\examples{
library(FWDselect)
data(diabetes)
x = diabetes[ ,2:11]
y = diabetes[ ,1]
obj2 = qselection(x, y, qvector = c(1:9), method = "lm", criterion = "variance", cluster = FALSE)
obj2
}
\author{
Marta Sestelo, Nora M. Villanueva and Javier Roca-Pardinas.
}
\seealso{
\code{\link{selection}}.
}

